Asset ID tab QoL, Voice ID parsing, code refactoring
Added a textbox in the Asset ID tab for what character is using the currently highlighted tab.
Added code for Voice ID reading and writing (no editor yet).

Code refactoring, got rid of redundant/unnecessary for loops, replaced if/else if conditionals with a switch statement as they will have more conditions in the future as the tool develops.